requirements.txt
appdirs==1.4.4
asgiref==3.7.2
atomicwrites==1.4.0
backports.weakref==1.0.post1
bkcharts==0.2
black==19.10b0
brotlipy==0.7.0
certifi==2021.10.8
constantly==15.1.0
cssselect==1.1.0
cytoolz==0.11.0
daal4py==2021.5.0
distlib==0.3.7
Django==4.2.5
django-allauth==0.58.1
django-cors-headers==4.2.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
filelock==3.12.4
fonttools==4.25.0
google-oauth2-tool==0.0.3
inflection==0.5.1
llvmlite==0.38.0
mccabe==0.6.1
mkl-fft==1.3.1
mkl-service==2.4.0
mpmath==1.2.1
multitasking==0.0.11
munkres==1.1.4
mypy-extensions==0.4.3
pandas==1.4.2
pathspec==0.7.0
patsy==0.5.2
pep8==1.7.1
Pillow==9.0.1
pipenv==2023.9.8
platformdirs==3.10.0
protobuf==3.19.1
psycopg2-binary==2.9.9
pyasn1-modules==0.2.8
pycosat==0.6.3
pycurl==7.44.1
PyDispatcher==2.0.5
pyls-spyder==0.4.0
pytest==7.1.1
python-constraint==1.4.0
python-decouple==3.8
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.2.4
pytz==2021.3
PyYAML==6.0
queuelib==1.5.0
sqlparse==0.4.4
tabulate==0.8.9
virtualenv==20.24.5
webencodings==0.5.1
wooldridge==0.4.4
xlwings==0.24.9
yfinance==0.1.74
zict==2.0.0
